requirements.txt
argcomplete==0.8.1
beautifulsoup4==4.4.1
coverage==4.4.1
dj-database-url==0.4.2
Django==1.11.5
django-filter==1.1.0
django-import-export==0.5.1
django-nose==1.4.5
djangorestframework==3.7.1
django-storages==1.6.5
django-jenkins
pep8
pylint
pyflakes
html5lib==0.999
httpie==0.9.9
httplib2==0.9.1
idna==2.6
mysqlclient==1.3.12
nose==1.3.7
Pillow==4.2.1
python-decouple==3.1
PyYAML==3.12
requests==2.18.4
tablib==0.12.1
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
XlsxWriter==0.7.3
gunicorn==19.7.1
python-decouple==3.1
